Extract shared HTTP logic into doHTTP
Both CustomerIO.request() and APIClient.doRequest() duplicated the
same marshal → create request → set headers → execute → read body
pipeline. Extract doHTTP() that accepts an auth callback, then have
both clients delegate to it. Auth (Basic vs Bearer) and error
handling remain caller-specific.

Also fixes a minor bug: CustomerIO.request() was not setting the
User-Agent header on DELETE requests (nil body path).
