+++ api.go
@@ -1,10 +1,7 @@
 package customerio
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
-	"io"
 	"net/http"
 )
 
@@ -37,36 +34,8 @@ func NewAPIClient(key string, opts ...Option) *APIClient {
 	return client
 }
 
-// doRequest returns the raw response body and status code. Callers
-// must check the status code and handle errors; unlike CustomerIO.request(),
-// this method does not return an error for non-200 responses.
 func (c *APIClient) doRequest(ctx context.Context, verb, requestPath string, body any) ([]byte, int, error) {
-	b, err := json.Marshal(body)
-	if err != nil {
-		return nil, 0, err
-	}
-
-	req, err := http.NewRequestWithContext(ctx, verb, c.URL+requestPath, bytes.NewBuffer(b))
-	if err != nil {
-		return nil, 0, err
-	}
-
-	req.Header.Set("Authorization", "Bearer "+c.Key)
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Add("User-Agent", c.UserAgent)
-
-	resp, err := c.Client.Do(req)
-	if err != nil {
-		return nil, 0, err
-	}
-	defer func() {
-		_ = resp.Body.Close()
-	}()
-
-	respBody, err := io.ReadAll(io.LimitReader(resp.Body, 1<<20))
-	if err != nil {
-		return nil, 0, err
-	}
-
-	return respBody, resp.StatusCode, nil
+	return doHTTP(ctx, c.Client, verb, c.URL+requestPath, c.UserAgent, body, func(req *http.Request) {
+		req.Header.Set("Authorization", "Bearer "+c.Key)
+	})
 }
