+++ http.go
@@ -1,6 +1,10 @@
 package customerio
 
 import (
+	"bytes"
+	"context"
+	"encoding/json"
+	"io"
 	"net/http"
 	"time"
 )
@@ -15,6 +19,47 @@ func newDefaultHTTPClient() *http.Client {
 		Timeout:   DefaultHTTPTimeout,
 		Transport: newDefaultTransport(),
 	}
+}
+
+// doHTTP is the shared HTTP execution path for both CustomerIO (Track) and
+// APIClient (App API). Auth header injection is caller-supplied via setAuth.
+func doHTTP(ctx context.Context, client HTTPClient, method, url, userAgent string, body any, setAuth func(*http.Request)) ([]byte, int, error) {
+	var req *http.Request
+	if body != nil {
+		b, err := json.Marshal(body)
+		if err != nil {
+			return nil, 0, err
+		}
+		req, err = http.NewRequestWithContext(ctx, method, url, bytes.NewBuffer(b))
+		if err != nil {
+			return nil, 0, err
+		}
+		req.Header.Set("Content-Type", "application/json")
+	} else {
+		var err error
+		req, err = http.NewRequestWithContext(ctx, method, url, nil)
+		if err != nil {
+			return nil, 0, err
+		}
+	}
+
+	req.Header.Set("User-Agent", userAgent)
+	setAuth(req)
+
+	resp, err := client.Do(req)
+	if err != nil {
+		return nil, 0, err
+	}
+	defer func() {
+		_ = resp.Body.Close()
+	}()
+
+	respBody, err := io.ReadAll(io.LimitReader(resp.Body, 1<<20))
+	if err != nil {
+		return nil, 0, err
+	}
+
+	return respBody, resp.StatusCode, nil
 }
 
 func newDefaultTransport() http.RoundTripper {
